Move config class to configs namespace.
