+++ osu.Game.Rulesets.Karaoke/Edit/KaraokeBeatmapVerifier.cs
@@ -7,6 +7,7 @@ using osu.Game.Beatmaps;
 using osu.Game.Rulesets.Edit;
 using osu.Game.Rulesets.Edit.Checks.Components;
 using osu.Game.Rulesets.Karaoke.Edit.Checks;
+using osu.Game.Rulesets.Karaoke.Edit.Checks.Configs;
 
 namespace osu.Game.Rulesets.Karaoke.Edit
 {
@@ -15,7 +16,7 @@ namespace osu.Game.Rulesets.Karaoke.Edit
         private readonly List<ICheck> checks = new List<ICheck>
         {
             // todo : implement config apply.
-            new CheckInvalidLyrics(new LyricCheckerConfig())
+            new CheckInvalidLyrics(new LyricCheckerConfig().CreateDefaultConfig())
         };
 
         public IEnumerable<Issue> Run(IBeatmap beatmap) => checks.SelectMany(check => check.Run(beatmap));
